requirements.txt
pytoniq>=0.1.38
pytoniq-core>=0.1.38
python-dotenv>=1.0.0
anthropic>=0.25.0
aiohttp>=3.9.0
